Lag støtte for double/desimaltall

+++ modell/src/main/kotlin/no/nav/dagpenger/behandling/hendelser/SøknadInnsendtHendelse.kt
@@ -26,7 +26,7 @@ class SøknadInnsendtHendelse(private val søknadId: UUID, private val journalpo
             fastsettelse<String>("Rettighetstype")
         }
         steg {
-            fastsettelse<Int>("Fastsatt vanlig arbeidstid") {
+            fastsettelse<Double>("Fastsatt vanlig arbeidstid") {
                 avhengerAv(virkningsdato)
             }
         }
